+++ src/main/java/com/nbicc/gywlw/Model/GywlwUser.java
@@ -32,7 +32,7 @@ public class GywlwUser {
     private Boolean delMark;
 
     //记录用户以什么身份登录的,普通用户0，设备厂商1；
-    private Byte userType;
+    private Integer userType;
 
     private String content;
 
@@ -44,11 +44,11 @@ public class GywlwUser {
         this.content = content;
     }
 
-    public Byte getUserType() {
+    public Integer getUserType() {
         return userType;
     }
 
-    public void setUserType(Byte userType) {
+    public void setUserType(Integer userType) {
         this.userType = userType;
     }
