+++ src/main/java/com/nbicc/gywlw/Model/LoginTicket.java
@@ -11,13 +11,13 @@ public class LoginTicket {
     private Date expired;
     private Integer status;//0有效，1无效；
     private String ticket;
-    private Byte userType;
+    private Integer userType;
 
-    public Byte getUserType() {
+    public Integer getUserType() {
         return userType;
     }
 
-    public void setUserType(Byte userType) {
+    public void setUserType(Integer userType) {
         this.userType = userType;
     }
